backend/requirements.txt
PyPDF2
openpyxl
reportlab
pdf2image
pandas
numpy
pydantic[email]
email-validator
mysql-connector-python
PyJWT
fastapi
uvicorn
sqlalchemy>=1.4
pymysql
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart
pydantic-settings
alembic
httpx
pytest
pytest-asyncio
pytest-cov
